Refactored service methods in object files
